s3 storage delete single file.

+++ routes/categoryRoutes.js
@@ -70,42 +70,44 @@ router.route('/delete-category').delete(protected, async (req, res) => {
     if (accountType === 'admin') {
         try {
             const category = await Category.findByIdAndDelete({ _id });
-            for (const business of category.businesses) {
-                const deletedBusiness = await Business.findByIdAndDelete({ _id: business._id });
-                // Delete address
-                await Address.findByIdAndDelete({ _id: deletedBusiness.businessAddress });
-                // Delete likes
-                await Like.findByIdAndDelete({ _id: deletedBusiness.likes});
-                // Delete comments
-                for (const commentID of deletedBusiness.comments) {
-                    const comment = await Comment.findByIdAndDelete({ _id: commentID });
-                    // Delete associated likes
-                    await Like.findByIdAndDelete({ _id: comment.likes });
-                }
-                // Delete reviews
-                for (const reviewID of deletedBusiness.reviews) {
-                    const review = await Review.findByIdAndDelete({ _id: reviewID });
-                    // Delete associated likes
-                    await Like.findByIdAndDelete({ _id: review.likes });
-                }
-                // Delete offers
-                for (const offerID of deletedBusiness.offers) {
-                    const offer = await Offer.findByIdAndDelete({ _id: offerID });
-                    // Delete associated likes
-                    await Like.findByIdAndDelete({ _id: offer.likes });
-                }
-                // Delete business images 
-                // Code goes here
+            if (category.businessCount > 0) {
+                for (const business of category.businesses) {
+                    const deletedBusiness = await Business.findByIdAndDelete({ _id: business._id });
+                    // Delete address
+                    await Address.findByIdAndDelete({ _id: deletedBusiness.businessAddress });
+                    // Delete likes
+                    await Like.findByIdAndDelete({ _id: deletedBusiness.likes});
+                    // Delete comments
+                    for (const commentID of deletedBusiness.comments) {
+                        const comment = await Comment.findByIdAndDelete({ _id: commentID });
+                        // Delete associated likes
+                        await Like.findByIdAndDelete({ _id: comment.likes });
+                    }
+                    // Delete reviews
+                    for (const reviewID of deletedBusiness.reviews) {
+                        const review = await Review.findByIdAndDelete({ _id: reviewID });
+                        // Delete associated likes
+                        await Like.findByIdAndDelete({ _id: review.likes });
+                    }
+                    // Delete offers
+                    for (const offerID of deletedBusiness.offers) {
+                        const offer = await Offer.findByIdAndDelete({ _id: offerID });
+                        // Delete associated likes
+                        await Like.findByIdAndDelete({ _id: offer.likes });
+                    }
+                    // Delete business images 
+                    // Code goes here
 
-                // Remove user comments ObjectIds from array
-                await User.update(
-                    { },
-                    { $pull: { 
-                        comments: { $in: business.comments},
-                        reviews: { $in: business.reviews}
-                    }},
-                    { multi: true }
-                );
+                    // Remove user comments ObjectIds from array
+                    await User.updateMany(
+                        { },
+                        { $pull: { 
+                            comments: { $in: business.comments},
+                            reviews: { $in: business.reviews}
+                        }},
+                        { multi: true }
+                    );
+                }
             }
             res.status(200).json(category);
